requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
fonttools==4.58.1
h11==0.16.0
holidays==0.74
idna==3.10
importlib_resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
narwhals==1.41.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
prophet==1.1.7
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
stanio==0.5.1
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
